requirements.txt
mistralai==1.4.0
python-dotenv==1.0.1
requests==2.32.3
numpy==2.2.2
matplotlib==3.10.0